backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2

numpy
torch
transformers==4.44.2
datasets==2.21.0
accelerate==0.34.2

faiss-cpu
